REFACTOR: risky - move functions to file with usage
